Make gui/math3d classes use float rather than qreal
This corrects the mismatch between using floats for internal storage
and qreal in the API of QVector*D which leads to lots of implicit
casts between double and float.

This change also stops users from being surprised by the loss of
precision when using these classes on desktop platforms and removes
the need for the private constructors taking a dummy int as the final
argument.

The QMatrix4x4 and QQuaternion classes have been changed to use float
for their internal storage since these are meant to be used in
conjunction with the QVector*D classes. This is to prevent unexpected
loss of precision and to improve performance.

The on-disk format has also been changed from double to float thereby
reducing the storage required when streaming vectors and matrices. This
is potentially a large saving when working with complex 3D meshes etc.

This also has a significant performance improvement when passing
matrices to QOpenGLShaderProgram (and QGLShaderProgram) as we no
longer have to iterate and convert the data to floats. This is
an operation that could easily be needed many times per frame.

This change also opens the door for further optimisations of these
classes to be implemented by using SIMD intrinsics.

This needs to be applied in conjunction with

https://codereview.qt-project.org/#change,33548

Task-number: QTBUG-21035
Task-number: QTBUG-20661

+++ src/gui/math3d/qgenericmatrix.cpp
@@ -273,7 +273,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix2x2 type defines a convenient instantiation of the
-    QGenericMatrix template for 2 columns, 2 rows, and qreal as
+    QGenericMatrix template for 2 columns, 2 rows, and float as
     the element type.
 */
 
@@ -282,7 +282,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix2x3 type defines a convenient instantiation of the
-    QGenericMatrix template for 2 columns, 3 rows, and qreal as
+    QGenericMatrix template for 2 columns, 3 rows, and float as
     the element type.
 */
 
@@ -291,7 +291,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix2x4 type defines a convenient instantiation of the
-    QGenericMatrix template for 2 columns, 4 rows, and qreal as
+    QGenericMatrix template for 2 columns, 4 rows, and float as
     the element type.
 */
 
@@ -300,7 +300,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix3x2 type defines a convenient instantiation of the
-    QGenericMatrix template for 3 columns, 2 rows, and qreal as
+    QGenericMatrix template for 3 columns, 2 rows, and float as
     the element type.
 */
 
@@ -309,7 +309,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix3x3 type defines a convenient instantiation of the
-    QGenericMatrix template for 3 columns, 3 rows, and qreal as
+    QGenericMatrix template for 3 columns, 3 rows, and float as
     the element type.
 */
 
@@ -318,7 +318,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix3x4 type defines a convenient instantiation of the
-    QGenericMatrix template for 3 columns, 4 rows, and qreal as
+    QGenericMatrix template for 3 columns, 4 rows, and float as
     the element type.
 */
 
@@ -327,7 +327,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix4x2 type defines a convenient instantiation of the
-    QGenericMatrix template for 4 columns, 2 rows, and qreal as
+    QGenericMatrix template for 4 columns, 2 rows, and float as
     the element type.
 */
 
@@ -336,7 +336,7 @@ QT_BEGIN_NAMESPACE
     \relates QGenericMatrix
 
     The QMatrix4x3 type defines a convenient instantiation of the
-    QGenericMatrix template for 4 columns, 3 rows, and qreal as
+    QGenericMatrix template for 4 columns, 3 rows, and float as
     the element type.
 */
