+++ src/gui/math3d/qgenericmatrix.h
@@ -336,14 +336,14 @@ Q_OUTOFLINE_TEMPLATE void QGenericMatrix<N, M, T>::copyDataTo(T *values) const
 }
 
 // Define aliases for the useful variants of QGenericMatrix.
-typedef QGenericMatrix<2, 2, qreal> QMatrix2x2;
-typedef QGenericMatrix<2, 3, qreal> QMatrix2x3;
-typedef QGenericMatrix<2, 4, qreal> QMatrix2x4;
-typedef QGenericMatrix<3, 2, qreal> QMatrix3x2;
-typedef QGenericMatrix<3, 3, qreal> QMatrix3x3;
-typedef QGenericMatrix<3, 4, qreal> QMatrix3x4;
-typedef QGenericMatrix<4, 2, qreal> QMatrix4x2;
-typedef QGenericMatrix<4, 3, qreal> QMatrix4x3;
+typedef QGenericMatrix<2, 2, float> QMatrix2x2;
+typedef QGenericMatrix<2, 3, float> QMatrix2x3;
+typedef QGenericMatrix<2, 4, float> QMatrix2x4;
+typedef QGenericMatrix<3, 2, float> QMatrix3x2;
+typedef QGenericMatrix<3, 3, float> QMatrix3x3;
+typedef QGenericMatrix<3, 4, float> QMatrix3x4;
+typedef QGenericMatrix<4, 2, float> QMatrix4x2;
+typedef QGenericMatrix<4, 3, float> QMatrix4x3;
 
 #ifndef QT_NO_DEBUG_STREAM
